+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 - `git`: [v0.5.1](services/git/CHANGELOG.md#v051-2025-06-04)
     - **Bugfix:** Adjusted `UnmarshalJSON` function to use enum types and added tests for enums
 - `loadbalancer`: 
+  - [v1.4.0](services/loadbalancer/CHANGELOG.md#v140-2025-06-12)
+    - **Feature:** Add new field `DisableTargetSecurityGroupAssignment` in `LoadBalancer` Model
   - [v1.3.0](services/loadbalancer/CHANGELOG.md#v130-2025-06-10)
     - **Feature:** Add new field `TargetSecurityGroup` in `LoadBalancer` Model
   - [v1.2.1](services/loadbalancer/CHANGELOG.md#v121-2025-06-04)